fixed item count bug

+++ src/items/Weapon.java
@@ -3,7 +3,7 @@ package items;
 public class Weapon extends Item {
 	private double damage;
 
-	public Weapon(String uniqueName, String name, int value, int damage) {
+	public Weapon(String uniqueName, String name, int value, double damage) {
 		super(uniqueName, name, value);
 		this.damage = damage;
 	}
